Moved tests to __tests__ folder
tests moved:
- requireLogin
- authRoutes
- signupRoutes
